requirements.txt
# Sahayak – AI Financial Assistant
# Python 3.10+ required
# Install: pip install -r requirements.txt

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.28.0

# LangChain core + integrations
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.2.0

# Vector store
chromadb>=0.5.0

# Data validation
pydantic>=2.6.0

# Environment variable loading
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
